Add a basic GUI in Java
I want this to work for anyone. Ideally it would run in a browser but you need to submit the scores and the point of this is that the scores never leave your computer, and browsers cannot update the local filesystem. (There are some libraries that allow you to do it in JavaScript but they are mostly only supported in Chrome and are using deprecated APIs). Hence, Java.

However, I have not used the package struture because the rest of the project won't be in Java. There will be a command line and a graphing function.

This commit is a basic GUI that offers one question and options and prints your choices to the command line. Clearly that's a long way from where I want to be but it's a start.

This commit also ignores the compiled classes.