Create the file if it doesn't already exist
So if it doesn't exist it will be created and if it does exist it will be opened and the new results appended.

I will then add the name of the file to `.gitignore` but since I don't plan for it to remain called `file.txt` I will do that later.