Pull out Question into its own class
Unfortunately, this compiles but I cannot test it; I can't figure out how to run both files. Using Java on the command line and coding in Vim is much different to an IDE, it turns out...

This needs much tidying, including imports and package names, which I'll fix once I've figured out how to run it.