The packages were wrong
I had forgotten some things about package structure, i.e. that the files had to actually be in a folder with that package name, it's not enough just to logically group them. This is why the previous commit didn't work.

I've also renamed the package to `phqs` rather than `anna`, though I should really stick with Java conventions and put it in a directory called `src`.

Anyway, the good news is that the previous commit fixed all the problems I was seeing - the button groups are correctly grouped and displayed.