Make window scrollable
Previously, everything needed to fit into the window so the radio buttons were squashed together, etc; and it would become even worse when we add the next set of questions.

This makes it scrollable. It fills the screen and is very sluggish when scrolling with the mouse. I think in order to overcome that I need to implement Scrollable as I think the default is to scroll one pixel per mouse movement.